Summary card tidying
Inverting info icon, move award to own component
